Disable the SpawnerPage GPU dropdown at zero
